Move static constants of ValueParser to class ValueParsers
