+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/BooleanMultiProperty.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.rule.properties;
 
-import static net.sourceforge.pmd.lang.rule.properties.ValueParser.BOOLEAN_PARSER;
+import static net.sourceforge.pmd.lang.rule.properties.ValueParsers.BOOLEAN_PARSER;
 
 import java.util.Arrays;
 import java.util.List;
@@ -12,7 +12,6 @@ import java.util.Map;
 
 import net.sourceforge.pmd.PropertyDescriptorFactory;
 import net.sourceforge.pmd.PropertyDescriptorField;
-import net.sourceforge.pmd.lang.rule.properties.ValueParser.Companion;
 
 /**
  * Defines a property type that supports multiple Boolean values.
@@ -29,7 +28,7 @@ public final class BooleanMultiProperty extends AbstractMultiValueProperty<Boole
                 char delimiter = delimiterIn(valuesById);
                 return new BooleanMultiProperty(nameIn(valuesById),
                                                 descriptionIn(valuesById),
-                                                Companion.parsePrimitives(defaultValueIn(valuesById), delimiter, BOOLEAN_PARSER),
+                                                ValueParsers.parsePrimitives(defaultValueIn(valuesById), delimiter, BOOLEAN_PARSER),
                                                 0f,
                                                 isDefinedExternally);
             }
