+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/BooleanProperty.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.rule.properties;
 
-import static net.sourceforge.pmd.lang.rule.properties.ValueParser.BOOLEAN_PARSER;
+import static net.sourceforge.pmd.lang.rule.properties.ValueParsers.BOOLEAN_PARSER;
 
 import java.util.Map;
