+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/DoubleProperty.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.rule.properties;
 
-import static net.sourceforge.pmd.lang.rule.properties.ValueParser.DOUBLE_PARSER;
+import static net.sourceforge.pmd.lang.rule.properties.ValueParsers.DOUBLE_PARSER;
 
 import java.util.Map;
