+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/FloatProperty.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.rule.properties;
 
-import static net.sourceforge.pmd.lang.rule.properties.ValueParser.FLOAT_PARSER;
+import static net.sourceforge.pmd.lang.rule.properties.ValueParsers.FLOAT_PARSER;
 
 import java.util.Map;
