+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/IntegerProperty.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.rule.properties;
 
-import static net.sourceforge.pmd.lang.rule.properties.ValueParser.INTEGER_PARSER;
+import static net.sourceforge.pmd.lang.rule.properties.ValueParsers.INTEGER_PARSER;
 
 import java.util.Map;
