+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/MethodMultiProperty.java
@@ -4,7 +4,7 @@
 
 package net.sourceforge.pmd.lang.rule.properties;
 
-import static net.sourceforge.pmd.lang.rule.properties.ValueParser.METHOD_PARSER;
+import static net.sourceforge.pmd.lang.rule.properties.ValueParsers.METHOD_PARSER;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
@@ -107,7 +107,7 @@ public final class MethodMultiProperty extends AbstractMultiPackagedProperty<Met
 
 
     private static List<Method> methodsFrom(String valueString) {
-        return ValueParser.Companion.parsePrimitives(valueString, MULTI_VALUE_DELIMITER, METHOD_PARSER);
+        return ValueParsers.parsePrimitives(valueString, MULTI_VALUE_DELIMITER, METHOD_PARSER);
     }
