+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/MethodProperty.java
@@ -5,7 +5,7 @@
 package net.sourceforge.pmd.lang.rule.properties;
 
 
-import static net.sourceforge.pmd.lang.rule.properties.ValueParser.METHOD_PARSER;
+import static net.sourceforge.pmd.lang.rule.properties.ValueParsers.METHOD_PARSER;
 
 import java.lang.reflect.Method;
 import java.util.Collections;
