+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/properties/TypeMultiProperty.java
@@ -83,7 +83,7 @@ public final class TypeMultiProperty extends AbstractMultiPackagedProperty<Class
 
 
     private static List<Class> typesFrom(String valueString) {
-        return ValueParser.Companion.parsePrimitives(valueString, MULTI_VALUE_DELIMITER, ValueParser.CLASS_PARSER);
+        return ValueParsers.parsePrimitives(valueString, MULTI_VALUE_DELIMITER, ValueParsers.CLASS_PARSER);
     }
 
 
@@ -101,7 +101,7 @@ public final class TypeMultiProperty extends AbstractMultiPackagedProperty<Class
 
     @Override
     protected Class createFrom(String toParse) {
-        return ValueParser.CLASS_PARSER.valueOf(toParse);
+        return ValueParsers.CLASS_PARSER.valueOf(toParse);
     }
